Create README specific to this project:
create-react-app generates a README.md file with lots of details
about create-react-app that may be useful to keep around, but not
to display so prominently as the README, so renamed it and created
a new README file that references the original.

The new README details how to setup this project, what version of
NodeJS etc.