Change Mapbox CSS URLs:
We need to include some 3rd party CSS files so the map is correctly
styled. I got these from the map overlay with redux mapbox-react-example.
I then updated the mapbox-gl.css to the same version of the mapbox-gl
dependency installed and this removed the `Missing Mapbox GL JS CSS`
error statement that was being displayed, incorrect version problem.
I also changed the URL and version of the assembly CSS file to the
latest.