Factor Toggle button out into it's own feature module:
Previously the Toggle button was part of the Map feature module. I
have now moved the code into it's own module for better separation
of concern.

The code to handle the actions/reducer functionality has been moved
to the Toggle module as this is really it's concern, not the map.