Update Satellite type buttons:
The Satellite Type toggle buttons in the top-left has now been moved
to the top-right and offset from the Navigation control. I've added
the correct names for the buttons but the data behind is still from
the blog, I need real data values so we can switch between Satellite
types.